backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.3
scipy==1.11.4
scikit-learn==1.3.2
pandas==2.1.3
